+++ include/mgard-x/DataRefactoring/MultiDimension/Coefficient/GridProcessingKernel.hpp
@@ -2248,9 +2248,9 @@ public:
     tbz = R;
     tby = C;
     tbx = F;
-    gridz = ceil((float)total_thread_z / tbz);
-    gridy = ceil((float)total_thread_y / tby);
-    gridx = ceil((float)total_thread_x / tbx);
+    gridz = ceil((double)total_thread_z / tbz);
+    gridy = ceil((double)total_thread_y / tby);
+    gridx = ceil((double)total_thread_x / tbx);
     for (DIM d = 0; d < D_GLOBAL; d++) {
       if (D_LOCAL == 3 && d != curr_dim_f && d != curr_dim_c &&
           d != curr_dim_r) {
@@ -4623,9 +4623,9 @@ public:
     tbz = R;
     tby = C;
     tbx = F;
-    gridz = ceil((float)total_thread_z / tbz);
-    gridy = ceil((float)total_thread_y / tby);
-    gridx = ceil((float)total_thread_x / tbx);
+    gridz = ceil((double)total_thread_z / tbz);
+    gridy = ceil((double)total_thread_y / tby);
+    gridx = ceil((double)total_thread_x / tbx);
     for (DIM d = 0; d < D_GLOBAL; d++) {
       if (D_LOCAL == 3 && d != curr_dim_f && d != curr_dim_c &&
           d != curr_dim_r) {
